Version 1.0.1

+++ src/main/java/se/iths/lab2webservice/mappers/BookMapper.java
@@ -21,7 +21,7 @@ public class BookMapper {
     }
 
     public Book mapp(BookDto bookDto){
-        return new Book(bookDto.getIsbn13(), bookDto.getTitel(), bookDto.getSpråk(),
+        return new Book((int) bookDto.getIsbn13(), bookDto.getTitel(), bookDto.getSpråk(),
                 bookDto.getPris(), bookDto.getUtgivningsdatum(), bookDto.getKategori());
     }
